[code] Circle: Fix radius type, constructor

+++ code/Circle.hpp
@@ -7,13 +7,13 @@ class Circle
 {
   private:
    Point _center;
-   unsigned int _radius;
+   double _radius;
 
   public:
    Circle(const Point& a, const Point& b, const Point& c);
 
    Point center() const { return _center; }
-   const unsigned int& radius() const { return _radius; }
+   const double& radius() const { return _radius; }
 
    ~Circle();
 };
